+++ ingest/exercise/hevy.py
@@ -3,14 +3,12 @@ import json
 import logging
 import os
 from datetime import datetime
-from pathlib import Path
 from typing import Any
 
-import boto3
 import conf
 import httpx
-import workout_entities as we  # Import the domain objects and parsing functions
-from botocore.exceptions import NoCredentialsError
+import utils
+import workout_entities as we  # Domain models and parsing functions
 from dotenv import load_dotenv
 
 # Load environment variables from .env file
@@ -32,28 +30,8 @@ logger = logging.getLogger(__name__)
 
 # Constant for the last processed date file name (as a string)
 LAST_PROCESSED_FILE: str = "last_processed_date.txt"
-
-# Define maximum page size as a constant
+# Define maximum page size as a constant (set to 10)
 MAX_PAGE_SIZE: int = 10
-
-# --- File operations: Create Path instances within functions ---
-
-
-def read_last_processed_date(file_name: str) -> str:
-    file_path: Path = Path(file_name)
-    if file_path.exists():
-        return file_path.read_text().strip()
-    else:
-        logger.info(
-            "No last processed date found. Using default: %s", conf.start_ingest_date
-        )
-        return conf.start_ingest_date
-
-
-def write_last_processed_date(file_name: str, date_str: str) -> None:
-    file_path: Path = Path(file_name)
-    file_path.write_text(date_str)
-
 
 # --- API fetching functions ---
 
@@ -94,23 +72,11 @@ async def fetch_all_events(
         return events
 
 
-# --- AWS S3 Upload function ---
-
-
-def upload_to_s3(data: str, bucket: str, key: str) -> None:
-    s3 = boto3.client("s3", region_name=conf.aws_region)
-    try:
-        s3.put_object(Bucket=bucket, Key=key, Body=data)
-        logger.info("Successfully uploaded data to s3://%s/%s", bucket, key)
-    except NoCredentialsError:
-        logger.error("AWS credentials not available.")
-
-
 # --- Main asynchronous function ---
 
 
 async def main() -> None:
-    last_processed_date: str = read_last_processed_date(LAST_PROCESSED_FILE)
+    last_processed_date: str = utils.read_last_processed_date(LAST_PROCESSED_FILE)
     events: list[we.Event] = await fetch_all_events(last_processed_date)
 
     if events:
@@ -119,7 +85,7 @@ async def main() -> None:
                 "Workout ID: %s, Title: %s", event.workout.id, event.workout.title
             )
         latest_date: str = max(event.workout.updated_at for event in events)
-        write_last_processed_date(LAST_PROCESSED_FILE, latest_date)
+        utils.write_last_processed_date(LAST_PROCESSED_FILE, latest_date)
 
         workouts_data_str: str = json.dumps(
             [event.__dict__ for event in events], default=lambda o: o.__dict__
@@ -127,7 +93,7 @@ async def main() -> None:
         s3_key: str = (
             f"{conf.s3_key_prefix}{datetime.now().strftime('%Y-%m-%d')}_workouts.json"
         )
-        upload_to_s3(workouts_data_str, conf.s3_bucket, s3_key)
+        utils.upload_to_s3(workouts_data_str, conf.s3_bucket, s3_key)
     else:
         logger.info("No new workouts found.")
